fix: subscriber update
