+++ packages/pubsub/src/DataParser.ts
@@ -1,5 +1,5 @@
 export class DataParser {
-  public parse(data: Buffer): unknown {
+  public parse(data: string): unknown {
     const dateTimeRegex = /^(\d{4}-\d\d-\d\d([tT][\d:.]*)?)([zZ]|([+-])(\d\d):?(\d\d))?$/
     const dateTimeReviver = (_: string, value: unknown) => {
       if (typeof value === 'string' && dateTimeRegex.test(value)) {
